builder/linode: move commHost() out of orphaned ssh.go
