+++ builder/linode/builder.go
@@ -6,6 +6,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"log"
 
 	"github.com/hashicorp/packer/common"
 	"github.com/linode/linodego"
@@ -95,4 +96,19 @@ func (b *Builder) Run(ctx context.Context, ui packer.Ui, hook packer.Hook) (ret 
 	}
 
 	return artifact, nil
+}
+
+func commHost(host string) func(multistep.StateBag) (string, error) {
+	return func(state multistep.StateBag) (string, error) {
+		if host != "" {
+			log.Printf("Using ssh_host value: %s", host)
+			return host, nil
+		}
+
+		instance := state.Get("instance").(*linodego.Instance)
+		if len(instance.IPv4) == 0 {
+			return "", fmt.Errorf("Linode instance %d has no IPv4 addresses!", instance.ID)
+		}
+		return instance.IPv4[0].String(), nil
+	}
 }
